requirements.txt
asgiref==3.5.2
certifi==2022.6.15.2
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
python-dotenv==0.21.0
requests==2.28.1
smmap==5.0.0
sqlparse==0.4.2
stripe==4.1.0
tzdata==2022.2
urllib3==1.26.12
